WFLY-11991 (Martin Simka test change) use SessionImplementor in other methods than nullSafeGet/nullSafeSet
